requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
mysql-connector-python==9.4.0
pyTelegramBotAPI==4.29.1
urllib3==2.5.0
